renamed module and exported apiLogger by name also
