+++ src/logging/index.js
@@ -1,7 +1,8 @@
-import { createLogger } from "./log";
+import { createLogger } from "./logger";
 
-export default createLogger("api");
-
+export const apiLogger = createLogger("api");
 export const emailSchedulerLogger = createLogger("emailSchedulerLog");
 export const calendarsSyncLog = createLogger("calendarsSyncLog");
 export const calendarsEventsSyncLog = createLogger("calendarsEventsSyncLog");
+
+export default apiLogger;
